requirements.txt
# Requirements for Deep Researcher Agent v2.0
pdfplumber>=0.9.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.21.0
torch>=1.12.0

# LangGraph and LangChain dependencies  
langgraph>=0.0.40
langchain-core>=0.1.40
langchain-community>=0.0.30
langchain-google-genai>=1.0.0

# Configuration and utilities
python-dotenv>=1.0.0
asyncio

# Report generation dependencies
reportlab>=3.6.0
markdown2>=2.4.0

# CLI and Web Interface dependencies
rich>=13.0.0
streamlit>=1.28.0
pandas>=1.5.0